Add files via upload
Done a fair amount here
Matrices is the hard extension where it allows for matrices to be created and added and multiplied, also allows to multiply a matrix with a twisted int but can't get it working the other way around yet
Did another (I think) hard extension that I made up, creating newTwist is the same as creating twistedints. when trying to add or multiply a newTwist with another new twist will be asked to create own function for adding, this function can be as long and complex as desired including brackets, powers etc. The function can also be either reused when adding two newTwists again or can be overwritten. Can hold two functions at once one used by doing + and one used by doing *